requirements.txt
streamlit
requests
sentence-transformers
faiss-cpu
pypdf